requirements/base.txt
# Configuration
django-environ==0.4.0
django-secure==1.0.1



# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1


# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.3
redis>=2.10.0



# Your custom requirements go here
https://bitbucket.org/ad_m/django-bleach/get/4576f3309920.zip
django-versatileimagefield==1.5
https://github.com/ad-m/django-atom/archive/v0.11.0.zip
django_filter==0.13.0
cached_property==1.3.0
django-extra-views==0.7.1
django-tinymce==2.3.0
django-geojson==2.9.0
django-leaflet==0.18.1
